subpfx: check if origins differ between 2 consecutive views
Prior to this commit, subpfxs_diff() would consider a super- and
sub-prefix pair to be the same if they simply exist in both the
current and previous view.

However, Iliana says that we should also check that the origins
of the sub-prefix and the origins of the super-prefix remain the
same between the two views as well. If not, then we should end
any current sub-prefix event and initiate a new one based on the
latest observed origins.

To support this, I've had to modify the pfx2pfx map to also
store the origins for each prefix so we can compare them. The
origins must be copied when we insert into the map, as the
origins pointed to by the user data pointer may be overwritten when
the next view is processed.

I've also added a couple of extra methods for working with
pt_user_t structures:
  pt_user_equal() and pt_user_clone()