Fix seeking failures when using librdkafka for input
rd_kafka_seek() does not work as expected if used on a topic that
has not been consumed from yet (and is considered deprecated).

rd_kafka_assign() is the preferred method for setting the offset
on a topic before begining consumption.

This commit updates the pfx and peer topic consumers to use
rd_kafka_assign(), as well as other corresponding API methods to
construct and manage the rd_kafka_topic_partition_list that is
required by rd_kafka_assign().

There are still some TODOs here:
 * add CLI argument to set the consumer group rather than
   hard-coding it to "bgpview-dev"
 * update meta topic consumers to also stop using rd_kafka_seek().