pergeo-ipinfo: revert to standard ipmeta lookup behavior
Recent changes to libipmeta have made it feasible to load the
entire ipinfo dataset into memory, so we do not need the
enhancements introduced to work with the memcache-psql provider.

Another change is that the consumer now also requires a CSV
file that lists all of the known regions and their corresponding
FQIDs, so that it can properly initialize the counters for all
regions before processing a view.