IPInfo geovisibility: ignore prefixes without complete IPInfo data
Previously, we threw an error if the IPInfo data for a prefix did
not account for all addresses in the prefix -- this was intended
to catch bugs in our handling of the IPInfo data, but now we
have seen a situation where the IPInfo data does have legitimate
gaps.

Instead, we will now log and ignore that particular prefix and
keep processing the rest of the prefixes.