pergeovisibility: add support for tracking country-AS and region-AS metrics
(country, AS) and (region, AS) metrics allow us to track
regional signals for ASes that have a presence in multiple
regions / countries and therefore better detect AS-level
outages that are only seen in particular locations.

A whitelist is used to determine which couplets are tracked
by the consumer, as many combinations are either redundant or
do not have enough /24s to be worth tracking.